PH-7: moved getStyledAttributes method to the GoogleSignInButton class
