+++ app/src/main/java/com/androiddev/profilehub/utils/ViewExtension.kt
@@ -45,60 +45,3 @@ fun View.customSnackbar(
         }
     return snackbar
 }
-
-fun GoogleSignInButton.getStyledAttributes(context: Context, attrs: AttributeSet?): GoogleButtonStyle {
-
-    val typedArray =
-        context.obtainStyledAttributes(attrs, R.styleable.GoogleSignInButton)
-
-    val text = try {
-        typedArray.getStringOrThrow(R.styleable.GoogleSignInButton_text)
-    } catch (_: Exception) {
-        Log.w(
-            this::class.java.name,
-            "parse: customText not found, using default: $DEFAULT_TEXT_GOOGLE_BUTTON"
-        )
-        DEFAULT_TEXT_GOOGLE_BUTTON
-    }
-
-    val textAllCaps = try {
-        typedArray.getBooleanOrThrow(R.styleable.GoogleSignInButton_textAllCaps)
-
-    } catch (_: Exception) {
-        Log.w(
-            this::class.java.name,
-            "parse: customTextAllCaps not found, using default: $DEFAULT_TEXT_All_CAPS_GOOGLE_BUTTON"
-        )
-        DEFAULT_TEXT_All_CAPS_GOOGLE_BUTTON
-    }
-
-    val textSizePx = try {
-        typedArray.getDimensionOrThrow(R.styleable.GoogleSignInButton_textSize)
-    } catch (_: Exception) {
-        val textSize = DEFAULT_TEXT_SIZE_GOOGLE_BUTTON.spToPx(context)
-        Log.w(
-            this::class.java.name,
-            "parse: customTextSize not found, using default: $textSize px"
-        )
-        textSize
-    }
-
-    val letterSpacingEm = try {
-        typedArray.getFloatOrThrow(R.styleable.GoogleSignInButton_letterSpacing)
-    } catch (_: Exception) {
-        val letterSpacingPx = DEFAULT_LETTER_SPACING_GOOGLE_BUTTON
-        Log.w(
-            this::class.java.name,
-            "parse: customLetterSpacing not found, using default: $letterSpacingPx px"
-        )
-        letterSpacingPx
-    }
-
-    typedArray.recycle()
-
-    val letterSpacingPx = letterSpacingEm * textSizePx
-    return GoogleButtonStyle(
-        text = if (textAllCaps) text.uppercase() else text,
-        textSize = textSizePx,
-        letterSpacing = letterSpacingPx)
-}
